+++ nomad/job_endpoint_hooks.go
@@ -8,6 +8,23 @@ import (
 	"github.com/hashicorp/nomad/nomad/structs"
 )
 
+const (
+	// vaultConstraintLTarget is the lefthand side of the Vault constraint
+	// injected when Vault policies are used. If an existing constraint
+	// with this target exists it overrides the injected constraint.
+	vaultConstraintLTarget = "${attr.vault.version}"
+)
+
+var (
+	// vaultConstraint is the implicit constraint added to jobs requesting a
+	// Vault token
+	vaultConstraint = &structs.Constraint{
+		LTarget: vaultConstraintLTarget,
+		RTarget: ">= 0.6.1",
+		Operand: structs.ConstraintVersion,
+	}
+)
+
 type admissionController interface {
 	Name() string
 }
@@ -112,7 +129,7 @@ func (jobImpliedConstraints) Mutate(j *structs.Job) (*structs.Job, []error, erro
 		return j, nil, nil
 	}
 
-	// Add Vault constraints
+	// Add Vault constraints if no Vault constraint exists
 	for _, tg := range j.TaskGroups {
 		_, ok := policies[tg.Name]
 		if !ok {
@@ -122,7 +139,7 @@ func (jobImpliedConstraints) Mutate(j *structs.Job) (*structs.Job, []error, erro
 
 		found := false
 		for _, c := range tg.Constraints {
-			if c.Equals(vaultConstraint) {
+			if c.LTarget == vaultConstraintLTarget {
 				found = true
 				break
 			}